services/mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "MCP server for Cogni's memory system"
requires-python = ">=3.10"
license = { text = "Private" }
dependencies = [
    "cogni",
    "cogni-infra-core>=0.1.0",
    "mcp",
    "uv",
    "pydantic-settings",
]

[project.optional-dependencies]
dev = [
    "ruff",
]
test = [
    "pytest",
    "pytest-asyncio", 
    "pytest-mock",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv.sources]
cogni-infra-core = { workspace = true }

[tool.ruff]
target-version = "py310"
line-length = 100
